REFACTORING: reprojecting csv converter to having static generic method
