+++ src/main/java/com/example/waytogo/initialize/repository/CsvConverterGeneric.java
@@ -1,18 +1,17 @@
-package com.example.waytogo.initialize.service;
+package com.example.waytogo.initialize.repository;
 
 import com.opencsv.bean.CsvToBeanBuilder;
-import org.springframework.stereotype.Service;
+import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Repository;
 
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileReader;
 import java.util.List;
 
-@Service
-public class CsvServiceGeneric<T> implements CsvService<T> {
-
-    @Override
-    public List<T> convertCsvFileToCsvModel(File csvFile, Class<T> clazz) {
+@Component
+public class CsvConverterGeneric {
+    public static <T> List<T> convertCsvFileToCsvModel(File csvFile, Class<T> clazz) {
         try {
             return new CsvToBeanBuilder<T>(new FileReader(csvFile))
                     .withType(clazz)
