+++ src/main/java/com/example/waytogo/initialize/service/CsvServiceLoader.java
@@ -1,0 +1,9 @@
+package com.example.waytogo.initialize.service;
+
+import lombok.AllArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@AllArgsConstructor
+public class CsvServiceLoader {
+}
